Remove unused use statments, added type hint
