+++ ClassContent/Repository/ClassContentQueryBuilder.php
@@ -187,7 +187,7 @@ class ClassContentQueryBuilder extends QueryBuilder
 
     public function addTitleLike($expression)
     {
-        if (null != $expression) {
+        if (null !== $expression) {
             $this->andWhere(
                 $this->expr()->like(
                     'cc._label',
