+++ Resources/layouts/error/500.phtml
@@ -14,9 +14,9 @@
             <div class="row" id="bb5-mainLayoutRow">
                 <div class="span12">
                     <section role="main" class="section">
-                        <h1>Page non trouvée&hellip;</h1>
+                        <h1>Une erreur est survenue&hellip;</h1>
                         <div class="announcement-wrapper">
-                            <p>Soit la page recherchée n'est pas activée, soit elle n'existe pas.</p>
+                            <p>Contactez l'administrateur de ce site.</p>
                         </div>
                         <hr>
                         <div class="announcement-box">
