+++ Rest/Controller/ClassContentController.php
@@ -23,17 +23,16 @@
 
 namespace BackBee\Rest\Controller;
 
+use BackBee\AutoLoader\Exception\ClassNotFoundException;
+use BackBee\ClassContent\AbstractClassContent;
+use BackBee\Rest\Controller\Annotations as Rest;
+
 use Doctrine\ORM\Tools\Pagination\Paginator;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\Security\Core\Exception\AccessDeniedException;
-
-use BackBee\AutoLoader\Exception\ClassNotFoundException;
-use BackBee\ClassContent\AbstractClassContent;
-use BackBee\Rest\Controller\Annotations as Rest;
-use BackBee\Utils\File\File;
 
 /**
  * ClassContent API Controller.
